refactor(service): Split ConversionService into specialized services
