+++ src/main/java/edu/esipe/i3/ezipflix/videodispatcher/service/MediaService.java
@@ -1,0 +1,7 @@
+package edu.esipe.i3.ezipflix.videodispatcher.service;
+
+public interface MediaService {
+
+    boolean checkFileExists(String objectName);
+
+}
